replaced function of checking id to the new hook
